Update data cleaning functions
The forecast data set provided by the data expo and our cleaned version were too large to be uploaded to Github. We changed them into RDS files instead of .dat and .csv files so that they could fit on the repository. These files were updated to account for those changes.